+++ include/visionaray/detail/simple_buffer_rt.inl
@@ -91,7 +91,7 @@ void simple_buffer_rt<ColorFormat, DepthFormat>::end_frame()
 }
 
 template <pixel_format ColorFormat, pixel_format DepthFormat>
-void simple_buffer_rt<ColorFormat, DepthFormat>::resize(size_t w, size_t h)
+void simple_buffer_rt<ColorFormat, DepthFormat>::resize(int w, int h)
 {
     render_target::resize(w, h);
